test-requirements.txt
nose>=1.2.1
mock>=1.0.1
pyhamcrest>=1.6
pexpect>=4.2.1
ncclient>=0.5.2
pyeapi
